python/requirements.txt
kafka-python==2.0.2
opensimplex==0.3
